tests: Fix min image transfer granularity test
This test as written was guaranteed to be skipped. It would skip the
test if the first graphics queue family found had a
minImageTransferGranularity with a width, height, or depth of less
than 4. The spec guarantees that any graphics queue family's min
transfer granularity will be 1 for the width, height, and depth. The
rewritten test is now only skipped if no queue supporting copy
operations is found with a large enough granularity.

Also the usage flags on the src and dst image needed to be fixed.
Additionally, the extent of the copies dst and src images were changed
to always be multiples of the min transfer granularity.

Finally, I added a device profile that has a queue with a large enough
min transfer granularity to not skip this test.